+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # ChangeLog
 
 ## Next version
-* Completely refactored pipeline into a message bus
+* Converted a pipeline into a message bus
 * Simplified trooba API
+* Support for sync and async messages
+* Support for custom messages, request/response, publish/subscribe or a mix
 
 ## v0.3
 * Initial commit